python-service/requirements.txt
Flask==2.3.3
boto3==1.28.40
PyPDF2==3.0.1
Pillow==10.0.0
Werkzeug==2.3.7
